quilt.cpp: always cast numbers in test data helper
Address FIXME in helper whereby sometimes we were not casting the
number types.

+++ projects/test_models/boost_model/src/test_data/class_e_td.cpp
@@ -25,7 +25,7 @@
 namespace {
 
 int create_int(const unsigned int position) {
-    return position;
+    return static_cast<int>(position);
 }
 
 double create_double(const unsigned int position) {
