+++ projects/test_models/trivial_inheritance/src/test_data/parent_with_members_td.cpp
@@ -27,7 +27,7 @@
 namespace {
 
 int create_int(const unsigned int position) {
-    return position;
+    return static_cast<int>(position);
 }
 
 }
